refactor: remove constants module and inline components/utilities in Nuxt plugin
