+++ packages/motion/src/index.ts
@@ -7,7 +7,6 @@ export { useLayoutGroup } from './components/use-layout-group'
 export type { LayoutGroupProps } from './components/use-layout-group'
 export * from './components/context'
 export * from './value'
-export * from './constants'
 export * from './types'
 export * from './animation'
 export * from './utils'
